split resolver into separate files
