+++ prqlc/prql-compiler/src/semantic/module.rs
@@ -1,13 +1,15 @@
 use std::collections::{HashMap, HashSet};
 
 use anyhow::Result;
+use prqlc_ast::stmt::QueryDef;
+use prqlc_ast::Span;
 
-use crate::ir::pl::{Expr, Ident, TupleField, Ty};
+use crate::ir::pl::{Annotation, Expr, Ident, TupleField, Ty};
 use crate::Error;
 
-use super::{Lineage, LineageColumn};
+use super::{Lineage, LineageColumn, NS_MAIN, NS_QUERY_DEF};
 use super::{NS_INFER, NS_INFER_MODULE, NS_SELF};
-use crate::ir::decl::{Decl, DeclKind, Module, TableDecl, TableExpr};
+use crate::ir::decl::{Decl, DeclKind, Module, RootModule, TableDecl, TableExpr};
 
 impl Module {
     pub fn singleton<S: ToString>(name: S, entry: Decl) -> Module {
@@ -345,6 +347,99 @@ impl Module {
         }
 
         res
+    }
+}
+
+type HintAndSpan = (Option<String>, Option<Span>);
+
+impl RootModule {
+    pub(super) fn declare(
+        &mut self,
+        ident: Ident,
+        decl: DeclKind,
+        id: Option<usize>,
+        annotations: Vec<Annotation>,
+    ) -> Result<()> {
+        let existing = self.module.get(&ident);
+        if existing.is_some() {
+            return Err(Error::new_simple(format!("duplicate declarations of {ident}")).into());
+        }
+
+        let decl = Decl {
+            kind: decl,
+            declared_at: id,
+            order: 0,
+            annotations,
+        };
+        self.module.insert(ident, decl).unwrap();
+        Ok(())
+    }
+
+    /// Finds that main pipeline given a path to either main itself or its parent module.
+    /// Returns main expr and fq ident of the decl.
+    pub fn find_main_rel(&self, path: &[String]) -> Result<(&TableExpr, Ident), HintAndSpan> {
+        let (decl, ident) = self.find_main(path).map_err(|x| (x, None))?;
+
+        let span = decl
+            .declared_at
+            .and_then(|id| self.span_map.get(&id))
+            .cloned();
+
+        let decl = (decl.kind.as_table_decl())
+            .ok_or((Some(format!("{ident} is not a relational variable")), span))?;
+
+        Ok((&decl.expr, ident))
+    }
+
+    pub fn find_main(&self, path: &[String]) -> Result<(&Decl, Ident), Option<String>> {
+        let mut tried_idents = Vec::new();
+
+        // is path referencing the relational var directly?
+        if !path.is_empty() {
+            let ident = Ident::from_path(path.to_vec());
+            let decl = self.module.get(&ident);
+
+            if let Some(decl) = decl {
+                return Ok((decl, ident));
+            } else {
+                tried_idents.push(ident.to_string());
+            }
+        }
+
+        // is path referencing the parent module?
+        {
+            let mut path = path.to_vec();
+            path.push(NS_MAIN.to_string());
+
+            let ident = Ident::from_path(path);
+            let decl = self.module.get(&ident);
+
+            if let Some(decl) = decl {
+                return Ok((decl, ident));
+            } else {
+                tried_idents.push(ident.to_string());
+            }
+        }
+
+        Err(Some(format!(
+            "Expected a declaration at {}",
+            tried_idents.join(" or ")
+        )))
+    }
+
+    pub fn find_query_def(&self, main: &Ident) -> Option<&QueryDef> {
+        let ident = Ident {
+            path: main.path.clone(),
+            name: NS_QUERY_DEF.to_string(),
+        };
+
+        let decl = self.module.get(&ident)?;
+        decl.kind.as_query_def()
+    }
+
+    /// Finds all main pipelines.
+    pub fn find_mains(&self) -> Vec<Ident> {
+        self.module.find_by_suffix(NS_MAIN)
     }
 }
